TEST: Most functions operating
Machine boots up, homes, and will go to positions correctly. GO TO button needs work to allow the FSM to go to the move state while maintenance is active. FSM "Prep" State needs to be converted to Clamp state, with infrastructure to determine if "GO" has been pressed or it has been entered from the auto-cycle function. Drill position needs to be checked - the drill position seems to be at global 0,0 rather than establishing a local 0,0.