TEST: ZeroOffset getting reset
ZeroOffset in the 6k class is getting reset at some point in the procedure. This results in the default zero offset not getting loaded into the class such that conversions are not made to the global coordinate system. The standard offset that is being observed by the 6k. The standard offset being observed when switching from drill to tap is the missing coordinate transformation.